Fix wierd hook relying on state by moving logic back into Video.js
